Working 'Quests' class and 'SimpleAdapter' class
I have gotten the simpleAdapter class to accept my Quests class for
data, and figured out that ListView requires that the ListView XML tag
to have '@id/android:list' and not '@+id/list'. Moving on to make the
ListView items clickable, which probably means finding a way to pass on
the 'Quests' class to the display activity.