Set defaults in login screen, Implemented a listener in ListView
I have set default username and password for the login, since I do not
want to edit the manifest to change activity load order. I have got a
click listener running for my ListView so that I can start my next
activity, hopefully using and intent to pass the 'Quests' class on so
that I can populate the detailed view of the quest.